requirements.txt
Django==1.9
django-crispy-forms==1.6.0
django-filter==1.0.1
django-markdown-deux==1.0.5
django-pagedown==0.1.1
djangorestframework==3.5.3
djangorestframework-jwt==1.9.0
Markdown==2.6.7
markdown2==2.3.1
olefile==0.44
Pillow==3.1.0
PyJWT==1.4.2